Scroll_pictures10.html
This is an html code, embedded in JavaScript, for an accessible version of the Soundboard in written form. 
If using this code, please acknowledge Dr Marina Klemm and Dr Kathryn Jones from University of Auckland, Aotearoa New Zealand.
Do not use the images or audios provided here without authorization. 
Contact: marinaklemm@gmail.com